requirements.txt
pandas==1.3.4
nba_api==1.4.1
boto3
sqlalchemy
apache-airflow-providers-amazon